Fix checkstyle errors
Issue gh-1025
